Changes from second part of pilot project session.
Create CanvasContext to organize data and methods associated with an InkCanvas.

Use different color and size drawing attributes for each canvas.

Implement redo command using undo and redo stacks.

Start work on assigning ids to strokes in each context to enable removal of a stroke from one context to propagate to the other. The idea here is to eventually assign all strokes a context and stroke id as they are created, so the can be uniquely identified across contexts.